requirements.txt
appnope==0.1.0
args==0.1.0
clint==0.5.1
decorator==4.0.2
dj-database-url==0.3.0
Django==1.8.1
django-postgrespool==0.3.0
django-pylibmc==0.6.0
github3.py==0.9.4
gnureadline==6.3.3
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.1.0
libsaas==0.4
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
psycopg2==2.6
pycd==0.3.28
PyGithub==1.25.2
pylibmc==1.5.0
requests==2.7.0
simplegeneric==0.8.1
SQLAlchemy==1.0.4
traitlets==4.0.0
uritemplate.py==0.3.0
wheel==0.24.0
whitenoise==1.0.6
